+++ CHANGELOG.rst
@@ -8,9 +8,10 @@ lib_xua Change Log
   * ADDED:     Application note AN00247: Using lib_xua with lib_spdif (transmit) 
   * CHANGE:    I2S hardware resources no longer used globally and must be passed
     to XUA_AudioHub()
-  * CHANGE:    NO_USB define renamed to XUA_USB_EN
   * CHANGE:    XUA_AudioHub() no longer pars S/PDIF transmitter task
   * CHANGE:    Moved to lib_spdif (from module_spdif_tx & module_spdif_rx)
+  * CHANGE:    Define NUM_PDM_MICS renamed to XUA_NUM_PDM_MICS
+  * CHANGE:    Define NO_USB renamed to XUA_USB_EN
 
 0.1.2
 -----